parked.yml fixes

+++ data/tools.sh
@@ -50,9 +50,11 @@ function check_for_parked {
         fi
         ((count++))
     done < "$1"
-    [[ -f "parked_domains_${1}.tmp" ]] && cat "parked_domains_${1}.tmp" >> "$parked_domains_file"
-    rm "$1"
-    rm "parked_domains_${1}.tmp"
+    if [[ -f "parked_domains_${1}.tmp" ]]; then
+        cat "parked_domains_${1}.tmp" >> "$parked_domains_file"
+        rm "$1"
+        rm "parked_domains_${1}.tmp"
+    fi
 }
 
 [[ "$1" == 'format' ]] && format "$2"
